+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Increment `@folio/stripes` to `v5`, `react-router` to `v5.2`.
 * Create access to New fast add record template from Check out screen. Refs UICHKOUT-628.
 * Do not send proxy info when patron has a proxy but is acting as self. Fixes UICHKOUT-644.
+* Handle malformed timestamps. Refs UICHKOUT-649.
 
 ## [4.0.1](https://github.com/folio-org/ui-checkout/tree/v4.0.1) (2020-06-19)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v4.0.0...v4.0.1)